requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
alpaca-trade-api==3.2.0
aiohttp==3.13.3
pytest==7.4.3
httpx==0.25.1
bcrypt==4.1.2
PyJWT==2.8.0
